Добавлена функция подсчета времени и разбиение на модули
